\name{Iceland}
\alias{Iceland}
\docType{data}
\title{Iceland Female Mortality Data}
\description{Mortality data of women in the calendar year 2006 from Iceland. The data were obtained from the Humam Mortality Database. Only ages from 40 to 110 have been included.}
\usage{data(Iceland)}
\format{
   This data frame contains 71 rows and the following 2 columns.
  \describe{
    \item{\code{D}}{Death counts for women of ages between 40 and 110 during the calendar year 2006}
    \item{\code{E}}{"Person-years" lived in the female population during the year 2006 for each age-group (from 40 to 110).}
}}
\source{
Human Mortality Database. University of California, Berkeley (USA), and Max Planck Institute for Demographic Research (Germany). Available at www.mortality.org or www.humanmortality.de
  
%%Statistics Iceland (http://www.statice.is/) 
}
\references{
Gamiz, M.L., Mammen, E., Martinez-Miranda, M.D. and Nielsen, J.P.(2016). Double one-sided cross-validation of local linear hazards. Journal of the Royal Statistical Society B, 78, 755-779.

Spreeuw, J., Nielsen, J. P. and Jarner, S. F. (2013). A visual test of mixed hazard models,
SORT, 37, 149-??170.

}
\examples{
data(Iceland)
}
\keyword{datasets}
